+++ src/redux/consts/index.js
@@ -1,3 +1,5 @@
 // Consts
 
 export const CREATE_SURVEY = 'CREATE_SURVEY';
+export const UPDATE_SURVEY = 'UPDATE_SURVEY';
+export const GET_SURVEY_BY_ID = 'GET_SURVEY_BY_ID';