+++ src/redux/initialState/index.js
@@ -3,7 +3,7 @@ export const initialState = {
     {
       id: 1,
       title: 'React tools',
-      description: '"Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
+      description: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.',
       questions: [
         {
           id: 1,
@@ -67,4 +67,5 @@ export const initialState = {
     'OPEN',
     'SELECT',
   ],
+  survey: {},
 }